Wave 144: accessor inlining (bp_root, k_field, camp), btlMain twin reconstruction, k_field 001b9140 restore

+++ src/Event/Fcl/fclShop.c
@@ -2583,7 +2583,7 @@ u8 * FUN_003e8860(u32 param_1,u32 param_2)
 
   u64 uVar6;
 
-  long lVar7;
+  int lVar7;
 
   int iVar8;
